pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clinicalagent-bench"
version = "0.1.0"
description = "Evaluation framework for healthcare AI agents"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "Nithin Sarva"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "click>=8.1.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "litellm>=1.0.0",
    "faiss-cpu>=1.7.0",
    "numpy>=1.24.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
cab = "clinicalagent_bench.cli.main:cli"

[project.entry-points."pytest11"]
clinicalagent_bench = "clinicalagent_bench.pytest_plugin"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501", "F541", "UP042"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
